creating recipes and getting them back

+++ src/database/migrations/20210915010358_create_table_users.ts
@@ -1,27 +1,29 @@
 import { Knex } from 'knex';
 
 export async function up(knex: Knex): Promise<void> {
-  await knex.schema.hasTable('users').then(exists => {
+  await knex.schema.hasTable('users').then(async exists => {
     if (exists) return;
 
-    return knex.schema.createTable('users', table => {
-      table.string('id').primary();
+    await knex.schema.createTable('users', table => {
+      table.string('user_id').primary().unique();
 
-      table.string('name').notNullable();
-      table.string('username').unique().notNullable();
+      table.string('name', 16).notNullable();
+      table.string('username', 8).unique().notNullable();
       table.string('email').unique().notNullable();
-      table.string('password').notNullable();
-      table.string('gender').notNullable();
+      table.string('password', 16).notNullable();
+      table.integer('gender').notNullable();
       table.string('biography');
-      table.integer('avatar_id').notNullable();
-      table.integer('friends_count');
 
       table.timestamp('created_at').defaultTo(knex.fn.now());
-      table.timestamp('updated_at').defaultTo(knex.fn.now());
+      table.timestamp('updated_at').defaultTo(knex.fn.now()); // set when updating data
     });
   });
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.schema.dropTable('users');
+  await knex.schema.hasTable('users').then(async exists => {
+    if (!exists) return;
+
+    await knex.schema.dropTable('users');
+  });
 }
